Cards added and Button
Location and Popular Location Cards were added. Button that links webpage to Figma Prototype was added in homepage. The footer was changed from "row" to "footer" and an additonal class was added to footer section so that no duplicate CSS exists. There were no changes to appears to the footer or cards after changing the class names.